requirements.txt
torch
transformers
tiktoken
groq
streamlit
pypdfium2
python-docx
nltk
requests